Move Calculator container into /src/containers/ and transfer Calculator component logic into Calculator container.
