+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Calculator from "./components/Calculator";
+import Calculator from "./containers/Calculator";
 
 const App = () => <Calculator />;
